requirements.txt
argcomplete==1.9.3
boto3==1.9.206
botocore==1.12.206
certifi==2019.6.16
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-braces==1.13.0
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-filter==2.2.0
django-jet==1.0.8
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.7.1
django-tinymce==2.8.0
djangorestframework==3.10.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
oauthlib==3.1.0
Pillow==6.1.0
placebo==0.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.6.1
python-monkey-business==1.0.0
python-slugify==1.2.4
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.0
Unidecode==1.1.1
urllib3==1.25.3
Werkzeug==0.15.5
wsgi-request-logger==0.4.6
zappa==0.48.2
